# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: RustDB Read-only HTTP Shell API
  version: 1.0.0-beta.1
  summary: Versioned background-query API used by the RustDB remote CLI.
  description: |
    RustDB v1 exposes one read-only SQL statement per background Query. Committed
    Arrow batches may be polled while execution runs; JSON/NDJSON pages become
    readable only after successful completion. Every `/v1`
    operation requires TLS and a principal Bearer Token; authentication runs
    before request-body or query-parameter parsing. Query-role principals can
    access only their own Query IDs, Admin principals can access every Query,
    and Idempotency Keys are scoped to the submitting principal. This specification deliberately
    excludes write SQL, Sessions, source administration, uploads, and a query
    listing operation.
servers:
  - url: https://127.0.0.1:7400
    description: Default loopback listener
...
security:
  - bearerAuth: []
paths:
...
  /v1/info:
    get:
      tags: [Protocol]
      operationId: serverInfo
      summary: Negotiate protocol and server capabilities
      responses:
        "200":
          description: Authenticated server information.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/queries:
    post:
      tags: [Queries]
      operationId: submitQuery
      summary: Submit one read-only SQL statement
      description: |
        A valid envelope creates a background Query. Repeating the same decoded
        request envelope with the same Idempotency-Key returns the original
        Query. JSON whitespace and object-member order are not part of the
        comparison. Reusing the key with a different envelope returns 409.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "202":
          description: Query accepted or an existing idempotent Query returned.
          headers:
            Location:
              description: Relative status resource for this Query.
              schema:
                type: string
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryAccepted"
        "400":
          description: Malformed JSON, invalid SQL, or an invalid request value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The Idempotency-Key was already used with another body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "413":
          description: Request exceeds the 1 MiB request-body limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "415":
          description: The request Content-Type is not application/json.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: The JSON shape is invalid or the SQL is outside the read-only allowlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: |
            The engine or principal admission queue is full
            (`admission.queue_full`), or the Query reservation exceeds a
            configured admission layer (`admission.resource_limit`). Queue-full
            responses are safely retryable and include Retry-After.
          headers:
            Retry-After:
              description: Suggested delay in seconds.
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/queries/{query_id}:
    parameters:
      - $ref: "#/components/parameters/QueryId"
    get:
      tags: [Queries]
      operationId: getQuery
      summary: Read Query status and bounded metrics
      responses:
        "200":
          description: Current Query state.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/QueryNotFound"
    delete:
      tags: [Queries]
      operationId: deleteQuery
      summary: Delete a terminal Query and its retained result
      description: This operation is idempotent after a Query is terminal.
      responses:
        "204":
          description: The terminal Query was deleted or was already absent.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The Query is queued or running; cancel it first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/queries/{query_id}/cancel:
    parameters:
      - $ref: "#/components/parameters/QueryId"
    post:
      tags: [Queries]
      operationId: cancelQuery
      summary: Idempotently request Query cancellation
      responses:
        "202":
          description: Cancellation was signalled, or the Query was terminal.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/QueryNotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Opaque principal credential read from an imported Profile Token file.
        The server stores only its SHA-256 digest. Query and Admin roles are
        authorized independently after authentication.
  headers:
    RequestId:
      description: Server-generated identifier for logs and support.
      schema:
        type: string
        minLength: 16
...
  parameters:
    QueryId:
      name: query_id
      in: path
      required: true
      description: Unguessable Query identifier returned by submission.
      schema:
        type: string
        minLength: 32
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Caller-generated 16-128 character visible ASCII key (no spaces or
        control characters). Its lifetime is the lifetime of the resulting
        ephemeral Query.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: "^[!-~]{16,128}$"
  responses:
...
    Unauthorized:
      description: Missing or invalid Bearer Token.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            const: 'Bearer realm="rustdb"'
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    QueryNotFound:
      description: Query ID is unknown, expired, deleted, or owned by another principal.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ServerInfo:
      type: object
      additionalProperties: false
      required:
        - protocol_version
        - server_version
        - read_only
        - capabilities
      properties:
        protocol_version:
          type: string
          const: v1
        server_version:
          type: string
          examples: [1.0.0-beta.1]
        read_only:
          type: boolean
          const: true
        capabilities:
          type: array
          uniqueItems: true
          items:
            type: string
            enum:
              - background_queries
              - arrow_ipc_results
              - batch_sequence_resume
              - typed_parameters
              - cursor_pagination
              - offset_pagination
              - json_results
              - ndjson_results
    QueryRequest:
      type: object
      additionalProperties: false
      required: [sql]
      properties:
        sql:
          type: string
          minLength: 1
          maxLength: 1048576
          description: Exactly one allowed read-only SQL statement.
        parameters:
          type: array
          default: []
          items:
            $ref: "#/components/schemas/TypedParameter"
        timeout_ms:
          type: integer
          format: int64
          minimum: 1
          description: Optional timeout shorter than or equal to the server limit.
    TypedParameter:
      type: object
      additionalProperties: false
      required: [type, value]
      description: |
        A typed positional parameter. Use JSON null as the value of any listed
        type to create a typed NULL. Decimal128 requires `precision` and
        `scale`; Binary is Base64, Date32 is epoch days, and
        TimestampMicrosecond is microseconds since the Unix epoch. Integer and
        Decimal values are JSON numbers; Decimal also accepts an exact decimal
        string. Clients must preserve the original numeric token for exactness.
      properties:
        type:
          type: string
          enum:
            - boolean
            - int64
            - uint64
            - float64
            - decimal128
            - utf8
            - binary
            - date32
            - timestamp_microsecond
        value:
          description: "Value matching `type`, or JSON null for a typed NULL."
        precision:
          type: integer
          minimum: 1
          maximum: 38
        scale:
          type: integer
          minimum: 0
          maximum: 38
    QueryAccepted:
      type: object
      additionalProperties: false
      required: [query_id, state, replayed, status_url, results_url]
      properties:
        query_id:
          type: string
        state:
          $ref: "#/components/schemas/QueryState"
        replayed:
          type: boolean
        status_url:
          type: string
        results_url:
          type: string
    QueryState:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    QueryStatus:
      type: object
      additionalProperties: false
      required:
        - query_id
        - state
        - created_at_ms
      properties:
        query_id:
          type: string
        state:
          $ref: "#/components/schemas/QueryState"
        created_at_ms:
          type: integer
          format: int64
          minimum: 0
          description: Milliseconds since the Unix epoch.
        started_at_ms:
          type: integer
          format: int64
          minimum: 0
          description: Omitted until execution starts.
        finished_at_ms:
          type: integer
          format: int64
          minimum: 0
          description: Omitted until the Query is terminal.
        metrics:
          $ref: "#/components/schemas/QueryMetrics"
          description: Present after successful execution; otherwise omitted.
        error:
          $ref: "#/components/schemas/Problem"
          description: Present for failed or cancelled Queries; otherwise omitted.
    QueryMetrics:
      type: object
      additionalProperties: false
      required:
        - elapsed_ms
        - rows_scanned
        - rows_returned
        - bytes_scanned
        - peak_memory_bytes
        - spill_read_bytes
        - spill_write_bytes
      properties:
        elapsed_ms:
          type: integer
          format: int64
          minimum: 0
        rows_scanned:
          type: integer
          format: int64
          minimum: 0
        rows_returned:
          type: integer
          format: int64
          minimum: 0
        bytes_scanned:
          type: integer
          format: int64
          minimum: 0
        peak_memory_bytes:
          type: integer
          format: int64
          minimum: 0
        spill_read_bytes:
          type: integer
          format: int64
          minimum: 0
        spill_write_bytes:
          type: integer
          format: int64
          minimum: 0
...
    Problem:
      type: object
      additionalProperties: false
      required: [error, message, retry]
      properties:
        error:
          type: string
          pattern: "^[a-z][a-z0-9_]*(\\.[a-z][a-z0-9_]*)*$"
          examples: [query.not_found]
        message:
          type: string
        retry:
          type: string
          enum: [never, safe, reopen_required, outcome_unknown, unknown]
          description: Machine-readable retry safety; clients must not infer this from `message`.
        request_id:
          type: string
        query_id:
          type: string
        details:
          type: object
          additionalProperties: true